Battery Charge & Death

+++ Assets/Scripts/Utils/Events/GameEvent.cs
@@ -16,7 +16,7 @@ public class GameEvent : ScriptableObject
     public UnityAction<Door> OnTryToUnlockDoor;
     public UnityAction<ICollectable> OnInteractItem;
     
-    public UnityAction<int> OnBatteryAdded;
+    public UnityAction<float> OnBatteryAdded;
     public UnityAction OnFinishRecharge;
 
     public UnityAction<LevelData> OnLevelChange;
